Create Music
select the title of all the songs by the artist named 'Queen'.

THEN 

make a 'Pop' playlist. In preparation, select the name of all of the artists from the 'Pop' genre.

THEN 

add another query that will select the title of all the songs from the 'Pop' artists. It should use IN on a nested subquery that's based on your previous query.